Create Coffee_Maker
Created a simple 'coffee' machine that allows a user to choose between 3 coffee drinks. 

The drink will only be made if there is enough money inserted and there are enough resources available. 

If a selection is not available, the user will be prompted again to choose a drink or they can check available resources.

Refactoring will continue when possible.